requirements.txt
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
bs4